Replace CFLAGS removing CPU extensions with more robust flag
We now use -mgeneral-regs-only on kernel compilation, this makes
the compiler only use general purpose registers which is what
disabling extensions tried to accomplish.

Also remove -mno-red-zone as this is 64 bit only flag. This will be
added back once kernel supportes 64 bit. And even then there will be
separate CFLAGS for 32 bit and 64 bit architectures.

I also enable link time optimization as that does good job optimizing
between CUs.

I am not sure why -fPIC is added as a flag, as the kernel is not
meant to be relocatable. I will leave it since it might be doing
something I am overlooking.